+++ src/main/java/org/example/rgybackend/Repository/CrisisAuditingRepository.java
@@ -3,6 +3,6 @@ package org.example.rgybackend.Repository;
 import org.example.rgybackend.Entity.CrisisAuditing;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CrisisAuditingRepository extends JpaRepository<CrisisAuditing, Long> {
+public interface CrisisAuditingRepository extends JpaRepository<CrisisAuditing, Integer> {
     
 }
