+++ src/main/java/org/example/rgybackend/Service/CrisisService.java
@@ -5,7 +5,6 @@ import org.example.rgybackend.Model.CrisisModel;
 import java.util.List;
 
 public interface CrisisService {
-
     void saveCrisis(int crisisid);
 
     void deleteCrisis(int crisisid);
@@ -20,5 +19,5 @@ public interface CrisisService {
 
     List<CrisisModel> getCrisisByUser(String userid);
   
-    boolean updateStatus(Long crisisid, Long status);
+    boolean updateStatus(Integer crisisid, Long status);
 }
